file structure organized and guide comments added
